Added "delay" animation field
Also fixed bug in Tweening library that captured "original value" too
early (on animation start, rather than process-delay-expired).  This
meant that out-and-in-again animations would have their second halves
fail, because it was trying to tween from (animation-start) "in" to "in"
values again, instead of the "out" value the previous animation process
left it at.